+++ app/controllers/partdescriptions_controller.rb
@@ -4,10 +4,12 @@ class PartdescriptionsController < ApplicationController
   end
 
   def new
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @partdescription = Partdescription.new
   end
 
   def create
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     params[:partdescription][:part_id] = params[:part_id]
     @partdescription = Partdescription.new(partdescription_params)
     if @partdescription.save
@@ -19,6 +21,7 @@ class PartdescriptionsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     partdescription = Partdescription.find(params[:id])
     partdescription.destroy
     redirect_to device_product_category_part_partdescriptions_path(params[:device_id], params[:product_id], params[:category_id], params[:part_id])
@@ -26,10 +29,12 @@ class PartdescriptionsController < ApplicationController
   end
 
   def edit
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @partdescription = Partdescription.find(params[:id])
   end
 
   def update
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @partdescription = Partdescription.find(params[:id])
     if @partdescription.update(partdescription_params)
       redirect_to device_product_category_part_partdescriptions_path(params[:device_id], params[:product_id], params[:category_id], params[:part_id])
