+++ app/controllers/products_controller.rb
@@ -10,10 +10,12 @@ class ProductsController < ApplicationController
   end
 
   def new
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @product = Product.new
   end
 
   def create
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     params[:product][:device_id] = params[:device_id]
     
     @product = Product.new(product_params)
@@ -38,10 +40,12 @@ class ProductsController < ApplicationController
   end
 
   def edit
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @product = Product.find(params[:id])
   end
 
   def update
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     @product = Product.find(params[:id])
     if @product.update(product_params)
       redirect_to device_products_path(params[:device_id])
@@ -52,6 +56,7 @@ class ProductsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path, :alert => "Unauthorized"   unless current_user.manager?
     product = Product.find(params[:id])
     product.destroy
     redirect_to device_products_path(params[:device_id])
